pyproject.toml
[project]
name = "cdisplayagain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.0.0",
    "pytest-cov>=6.1.0",
    "pytest>=9.0.2",
    "rarfile>=4.2",
    "ruff>=0.11.2",
]

[tool.pytest.ini_options]
addopts = ["--cov=cdisplayagain", "--cov-fail-under=99"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["D"]
ignore = ["D203", "D213"]
